+++ src/pages/healthRecord/index.js
@@ -1,15 +1,22 @@
-import Header from "../../components/Header";
-import HealthRecordForm from "../../components/HealthRecordForm";
+import Header from '../../components/Header';
+import HealthRecordTable from '../../components/PatientRecordsTable';
+import { fetchJson } from '../../utils/HttpUtils';
 
-function HealthRecord() {
+function HealthRecord(props) {
   return (
     <div>
       <Header />
       <main className="max-w-screen-md mx-auto py-10">
-        <HealthRecordForm />
+        <HealthRecordTable records={props.records} />
       </main>
     </div>
   );
 }
 
 export default HealthRecord;
+
+export async function getServerSideProps({}) {
+    const id = process.env.NEXT_PUBLIC_USER_ID;
+    const records = await fetchJson(`/HealthRecord/get/byPatient?patientId=${id}`);
+    return { props: { records } };
+}
