+++ src/pages/prescription/index.js
@@ -22,5 +22,7 @@ export async function getServerSideProps() {
 
   const prescription = await fetchJson(`/Recipe/get/byPatientId?patientid=${id}`);
 
+  console.log(prescription);
+
   return { props: { prescription } };
 }
